Everything Works
Ok, so maybe there are still some issues with randomness not being
guaranteed at exceedingly large sizes. Not to mention that calling
`malloc` every single time a wall is added to the list of walls under
consideration is horribly unfancy. However, I don't really need this to
be the world's most efficient maze generator, so this is probably where
I'll leave this project off. I might come back at some point & add some
unit tests, but that's pretty much all I have left on my to-do list for
this project.

Signed-off-by: Buckley Ross <buckleyross42@gmail.com>